requirements.txt
django>=3.1,<3.2
uwsgi==2.0.19.1
psycopg2==2.8.5
#dj-database-url==0.5.0
whitenoise==5.2.0
psycopg2-binary>=2.8
pillow
django-crispy-forms
django-import-export
